Restore makecancellable to a local util
