+++ packages/block-editor/src/components/link-control/use-create-page.js
@@ -4,11 +4,6 @@
 import { __ } from '@wordpress/i18n';
 import { useEffect, useState, useRef } from '@wordpress/element';
 
-/**
- * Internal dependencies
- */
-import makeCancelable from './make-cancelable';
-
 export default function useCreatePage( handleCreatePage ) {
 	const cancelableCreateSuggestion = useRef();
 	const [ isCreatingPage, setIsCreatingPage ] = useState( false );
@@ -63,3 +58,30 @@ export default function useCreatePage( handleCreatePage ) {
 		errorMessage,
 	};
 }
+
+/**
+ * Creates a wrapper around a promise which allows it to be programmatically
+ * cancelled.
+ * See: https://reactjs.org/blog/2015/12/16/ismounted-antipattern.html
+ *
+ * @param {Promise} promise the Promise to make cancelable
+ */
+const makeCancelable = ( promise ) => {
+	let hasCanceled_ = false;
+
+	const wrappedPromise = new Promise( ( resolve, reject ) => {
+		promise.then(
+			( val ) =>
+				hasCanceled_ ? reject( { isCanceled: true } ) : resolve( val ),
+			( error ) =>
+				hasCanceled_ ? reject( { isCanceled: true } ) : reject( error )
+		);
+	} );
+
+	return {
+		promise: wrappedPromise,
+		cancel() {
+			hasCanceled_ = true;
+		},
+	};
+};
